fix(cfapi): CF_SYNC_REGISTRATION.StructSize must be ushort, not uint
Root cause of persistent E_INVALIDARG (0x80070057) from CfRegisterSyncRoot:
The StructSize field in CF_SYNC_REGISTRATION is declared as USHORT (2 bytes)
in cfapi.h, but our C# struct used uint (4 bytes). This caused:

1. Struct size inflated by 2 bytes (extra padding after ushort alignment)
2. All subsequent field offsets shifted by 2 bytes
3. Marshal.SizeOf returned wrong value → API rejected the struct

Fix: Changed StructSize from uint to ushort. Added explicit (ushort) cast.
This matches the native cfapi.h definition and all Cloud Mirror samples.

+++ src/NasSync.CfApi/CfSyncRootManager.cs
@@ -153,7 +153,7 @@ public sealed class CfSyncRootManager : IDisposable
         // Step 2: Build the native registration structure
         var nativeRegistration = new CfNativeTypes.CF_SYNC_REGISTRATION
         {
-            StructSize = (uint)Marshal.SizeOf<CfNativeTypes.CF_SYNC_REGISTRATION>(),
+            StructSize = (ushort)Marshal.SizeOf<CfNativeTypes.CF_SYNC_REGISTRATION>(),
             ProviderName = registrationInfo.DisplayName,
             ProviderVersion = registrationInfo.Version,
             HydrationPolicy = (CfNativeTypes.CF_HYDRATION_POLICY)registrationInfo.HydrationPolicy,
